add new url and fix bugs

+++ pages/api/deleteJob.js
@@ -1,19 +1,28 @@
 import dbConnect from "../../db/connect";
 import Job from "../../db/models/Jobs";
+import { authOptions } from "pages/api/auth/[...nextauth]";
+import { getServerSession } from "next-auth/next";
 
 export default async function handler(request, response) {
+	const session = await getServerSession(request, response, authOptions);
+
+	if (!session) {
+		response.status(401).json({ error: "Unauthorized" });
+		return;
+	}
+
 	await dbConnect();
 
-    if (request.method === "DELETE") {
-        const id = request.body.id;
-        try {
-          await Job.findByIdAndDelete(id);
-          response
-            .status(200)
-            .json({ status: `job with id: ${id} successfully deleted.` });
-        } catch (error) {
-          console.log(error);
-          response.status(400).json({ error: error.message });
-        }
-      }
+	if (request.method === "DELETE") {
+		const id = request.body.id;
+		try {
+			await Job.findByIdAndDelete(id);
+			response
+				.status(200)
+				.json({ status: `job with id: ${id} successfully deleted.` });
+		} catch (error) {
+			console.log(error);
+			response.status(400).json({ error: error.message });
+		}
+	}
 }
